Update infrastructure in docker-compose file. Creates new Controller adapter.

+++ app/src/main/kotlin/com/br/rinhadebackend2/crebito/models/TransacaoRequest.kt
@@ -1,7 +1,7 @@
 package com.br.rinhadebackend2.crebito.models
 
 data class TransacaoRequest(
-    val valor: Int,
+    val valor: Long,
     val tipo: String,
     val descricao: String
 )
